+++ config/webpackDevServer.config.js
@@ -13,7 +13,6 @@ const fs = require('fs');
 const protocol = process.env.HTTPS === 'true' ? 'https' : 'http';
 const host = process.env.HOST || '0.0.0.0';
 
-const createEnvSettingsFile = require('../src/build/env');
 module.exports = function(proxy, allowedHost, decoratedIndexHtml) {
   return {
     // WebpackDevServer 2.4.3 introduced a security fix that prevents remote
@@ -106,10 +105,6 @@ module.exports = function(proxy, allowedHost, decoratedIndexHtml) {
 
       app.get('/personopplysninger-api/local/cookie', (req, res) => {
         res.send("This has to be a default login page");
-      });
-
-      app.get('/static/js/settings.js', (req, res) => {
-        res.send(createEnvSettingsFile());
       });
 
       app.get('/mock-api/person-info.json', (req, res) => {
