+++ server.js
@@ -4,7 +4,6 @@ const express = require('express');
 const path = require('path');
 const mustacheExpress = require('mustache-express');
 const getDecorator = require('./src/build/decorator');
-const createEnvSettingsFile = require('./src/build/env');
 
 const server = express();
 
@@ -35,7 +34,6 @@ const startServer = (html) => {
 
   server.use('person/personopplysninger/mock-api', express.static(path.resolve(__dirname, 'src/mock-api')));
   server.use('person/personopplysninger/static/js', express.static(path.resolve(__dirname, 'build/static/js')));
-  server.get('/static/js/settings.js', (req, res) => res.send(createEnvSettingsFile()));
 
   server.use(
     'person/personopplysninger/static/css',
@@ -51,9 +49,6 @@ const startServer = (html) => {
     'person/personopplysninger/static/media',
     express.static(path.resolve(__dirname, 'build/static/media')),
   );
-
-  server.get('/health/isAlive', (req, res) => res.sendStatus(200));
-  server.get('/health/isReady', (req, res) => res.sendStatus(200));
 
   server.get(/^\/(?!.*static).*$/, (req, res) => {
     res.send(html);
