--- src/js/Config.js
@@ -1,3 +1,0 @@
-export default {
-  personopplysninger: window.personopplysningerSettings, // eslint-disable-line no-undef
-};
