+++ src/js/initialState.js
@@ -1,18 +1,32 @@
 export default {
-  adresseinfo: {
+  personalia: {
+    fornavn: '',
+    etternavn: '',
+    fnr: '',
+    kontonr: '',
+    tlfnr: null,
+    spraak: null,
+    epostadr: '',
+    personstatus: '',
+    statsborgerskap: '',
+    foedested: '',
+    sivilstand: '',
+    kjoenn: '',
+    datakilder: [{}],
+  },
+  adresser: {
     boadresse: {
       adresse: '',
       adressetillegg: '',
       bydel: '',
       datoFraOgMed: '',
-      kilde: '',
       kommune: '',
       land: '',
       matrikkeladresse: {
-        bruksnummer: '',
-        festenummer: '',
-        gaardsnummer: '',
-        undernummer: '',
+        bruksnummer: null,
+        festenummer: null,
+        gaardsnummer: null,
+        undernummer: null,
       },
       postnummer: '',
       poststed: '',
@@ -26,219 +40,13 @@ export default {
     geografiskTilknytning: {
       bydel: '',
       datoFraOgMed: '',
-      kilde: '',
       kommune: '',
       land: '',
     },
-    postadresse: {
-      adresse1: '',
-      adresse2: '',
-      adresse3: '',
-      datoFraOgMed: '',
-      kilde: '',
-      land: '',
-      postnummer: '',
-    },
-    prioritertAdresse: {
-      datoFraOgMed: '',
-      kilde: '',
-      type: '',
-    },
-    tilleggsadresse: {
-      adresse1: '',
-      adresse2: '',
-      adresse3: '',
-      bolignummer: '',
-      bydel: '',
-      datoFraOgMed: '',
-      gateKode: '',
-      husbokstav: '',
-      husnummer: '',
-      kilde: '',
-      kommunenummer: '',
-      postboksanlegg: '',
-      postboksnummer: '',
-      postnummer: '',
-    },
-    utenlandskAdresse: {
-      adresse1: '',
-      adresse2: '',
-      adresse3: '',
-      datoFraOgMed: '',
-      kilde: '',
-    },
+    postadresse: null,
+    prioritertAdresse: null,
+    tilleggsadresse: null,
+    utenlandskAdresse: null,
+    datakilder: [{}],
   },
-  barnListe: [
-    {
-      datoRegistert: '',
-      egenansatt: null,
-      fnr: '',
-      kilde: '',
-      spesreg: '',
-    },
-  ],
-  brukerbehovListe: [
-    {
-      behov: '',
-      datoFraOgMed: '',
-      kilde: '',
-    },
-  ],
-  datoFraOgMed: '',
-  doedsdato: {
-    dato: '',
-    datoFraOgMed: '',
-    kilde: '',
-  },
-  egenansatt: {
-    datoFraOgMed: '',
-    erEgenansatt: null,
-    kilde: '',
-  },
-  foedselsdato: '',
-  foedtIKommune: '',
-  foedtILand: '',
-  foreldreansvar: {
-    datoFraOgMed: '',
-    kilde: '',
-    type: '',
-  },
-  ident: '',
-  identtype: '',
-  innvandringUtvandring: {
-    innvandretDato: '',
-    innvandretKilde: '',
-    innvandretLand: '',
-    utvandretDato: '',
-    utvandretKilde: '',
-    utvandretLand: '',
-  },
-  kilde: '',
-  kjonn: '',
-  kontonummer: {
-    datoFraOgMed: '',
-    kilde: '',
-    nummer: '',
-  },
-  navn: {
-    datoFraOgMed: '',
-    forkortetNavn: '',
-    fornavn: '',
-    kilde: '',
-    mellomnavn: '',
-    slektsnavn: '',
-    slektsnavnUgift: '',
-  },
-  oppholdstillatelse: {
-    datoFraOgMed: '',
-    kilde: '',
-    type: '',
-  },
-  relasjon: {
-    datoFraOgMed: '',
-    egenansatt: null,
-    fnr: '',
-    kilde: '',
-    spesreg: '',
-    type: '',
-  },
-  samboer: {
-    datoFraOgMed: '',
-    fnr: '',
-    kilde: '',
-  },
-  sivilstand: {
-    datoFraOgMed: '',
-    kilde: '',
-    kode: '',
-  },
-  spesreg: {
-    datoFraOgMed: '',
-    kilde: '',
-    kode: '',
-  },
-  spraak: {
-    datoFraOgMed: '',
-    kilde: '',
-    kode: '',
-  },
-  statsborgerskap: {
-    datoFraOgMed: '',
-    kilde: '',
-    kode: '',
-  },
-  status: {
-    datoFraOgMed: '',
-    kilde: '',
-    kode: '',
-  },
-  telefon: {
-    jobb: '',
-    jobbDatoRegistrert: '',
-    jobbKilde: '',
-    mobil: '',
-    mobilDatoRegistrert: '',
-    mobilKilde: '',
-    privat: '',
-    privatDatoRegistrert: '',
-    privatKilde: '',
-  },
-  tiltak: {
-    datoFraOgMed: '',
-    datoTil: '',
-    kilde: '',
-    type: '',
-  },
-  utenlandsinfoList: [
-    {
-      datoFraOgMed: '',
-      familienavnFodt: '',
-      farsFamilenavn: '',
-      farsFornavn: '',
-      foedested: '',
-      fornavnFodt: '',
-      idOff: '',
-      institusjon: '',
-      institusjonNavn: '',
-      kilde: '',
-      kildePin: '',
-      land: '',
-      morsFamilenavn: '',
-      morsFornavn: '',
-      nasjonalId: '',
-      nasjonalitet: '',
-      sedRef: '',
-      sektor: '',
-    },
-  ],
-  utenlandskBank: {
-    adresse1: '',
-    adresse2: '',
-    adresse3: '',
-    bankkode: '',
-    banknavn: '',
-    datoFraOgMed: '',
-    iban: '',
-    kilde: '',
-    kontonummer: '',
-    land: '',
-    swiftkode: '',
-    valuta: '',
-  },
-  vergemaalListe: [
-    {
-      aaksId: '',
-      datoFraOgMed: '',
-      egenansatt: null,
-      embete: '',
-      fnr: '',
-      id: '',
-      kilde: '',
-      mandattype: '',
-      sakstype: '',
-      spesreg: '',
-      type: '',
-      vedtaksdato: '',
-    },
-  ],
 };
